// src/types/apps/courses.d.ts
export interface ICourses {
  [key: string]: any
  first_name?: string
  last_name?: string
  phone?: string
  country?: string
  description?: string
  stripe_customerId?: string
  email_status?: string
  id?: string
  name?: string
  channelId?: string
  courseCompletePercentage?: number
  totalVideosCompletedCount?: number
  totalVideosCount?: number
}